+++ test/pretty_expressive.ml
@@ -5,12 +5,12 @@ let print_doc_choice (w : int) =
   let module P = Printer.Make (val cf) in
   let open P in
 
-  let d = text "while (true) {" <>
+  let d = text "while (true) {" ^^
           nest 4
-            (nl <> text "f();" <> nl <> text "if (done())" <>
+            (nl ^^ text "f();" ^^ nl ^^ text "if (done())" ^^
              (let exit_d = text "exit();" in
-              (space <> exit_d) <|> nest 4 (nl <> exit_d))) <>
-          nl <> text "}"
+              (space ^^ exit_d) <|> nest 4 (nl ^^ exit_d))) ^^
+          nl ^^ text "}"
   in
   pretty_print d
 
@@ -19,11 +19,11 @@ let print_doc_group (w : int) =
   let module P = Printer.Make (val cf) in
   let open P in
 
-  let d = text "while (true) {" <>
+  let d = text "while (true) {" ^^
           nest 4
-            (nl <> text "f();" <> nl <> text "if (done())" <>
-             group (nest 4 (nl <> text "exit();"))) <>
-          nl <> text "}"
+            (nl ^^ text "f();" ^^ nl ^^ text "if (done())" ^^
+             group (nest 4 (nl ^^ text "exit();"))) ^^
+          nl ^^ text "}"
   in
   pretty_print d