+++ CHANGES.md
@@ -1,3 +1,10 @@
+## 0.2 (2023-11-28)
+
+* Change `<>` to `^^` to avoid shadowing the not equal operator 
+  (thanks to @EmileTrotignon who made the suggestion in 
+  [the OCaml forum](https://discuss.ocaml.org/t/ann-first-release-of-pretty-expressive-a-pretty-expressive-printer/13516/2)).
+* Improve documentation
+
 ## 0.1 (2023-11-26)
 
 * Initial release